+++ package.json
@@ -55,6 +55,6 @@
     "ts-jest": "^23.0.1",
     "ts-node": "^8.3.0",
     "tslint": "^5.18.0",
-    "typescript": "^3.7.3"
+    "typescript": "^3.8.2"
   }
 }
